requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.22.0
Pillow==9.3.0
pyyaml==6.0
tqdm==4.64.0
imageio==2.26.0
